Revert "fix: one more attempt"
This reverts commit 0f3b3642

+++ core/network/impl/peer_manager_impl.cpp
@@ -112,7 +112,7 @@ namespace kagome::network {
     }
   }
 
-  uint32_t PeerManagerImpl::activePeersNumber() const {
+  size_t PeerManagerImpl::activePeersNumber() const {
     return active_peers_.size();
   }
