+++ core/network/impl/peer_manager_impl.hpp
@@ -58,7 +58,7 @@ namespace kagome::network {
     void connectToPeer(const PeerInfo &peer_info) override;
 
     /** @see PeerManager::activePeersNumber */
-    uint32_t activePeersNumber() const override;
+    size_t activePeersNumber() const override;
 
     /** @see PeerManager::forEachPeer */
     void forEachPeer(std::function<void(const PeerId &)> func) const override;
